requirements.txt
wsaccel
websocket-client
jwt
pandas
pprint
